+++ crates/cxx-qt-gen/src/generator/cpp/method.rs
@@ -3,11 +3,15 @@
 //
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-use crate::generator::cpp::get_cpp_params;
 use crate::{
-    generator::cpp::{
-        fragment::{CppFragment, CppNamedType},
-        qobject::GeneratedCppQObjectBlocks,
+    generator::{
+        cfg::try_eval_attributes,
+        cpp::{
+            fragment::{CppFragment, CppNamedType},
+            get_cpp_params,
+            qobject::GeneratedCppQObjectBlocks,
+            GeneratedOpt,
+        },
     },
     naming::cpp::{syn_return_type_to_cpp_except, syn_type_to_cpp_return_type},
     naming::TypeNames,
@@ -18,9 +22,15 @@ use syn::Result;
 pub fn generate_cpp_methods(
     invokables: &Vec<&ParsedMethod>,
     type_names: &TypeNames,
+    opt: &GeneratedOpt,
 ) -> Result<GeneratedCppQObjectBlocks> {
     let mut generated = GeneratedCppQObjectBlocks::default();
     for &invokable in invokables {
+        // Skip if the cfg attributes are not resolved to true
+        if !try_eval_attributes(opt.cfg_evaluator.as_ref(), &invokable.cfgs)? {
+            continue;
+        }
+
         let return_cxx_ty = syn_type_to_cpp_return_type(&invokable.method.sig.output, type_names)?;
 
         let parameters: Vec<CppNamedType> = get_cpp_params(&invokable.method, type_names)?;
@@ -77,9 +87,28 @@ mod tests {
     use super::*;
 
     use crate::generator::cpp::property::tests::require_header;
+    use crate::tests::CfgEvaluatorTest;
     use pretty_assertions::assert_str_eq;
     use std::collections::HashSet;
     use syn::{parse_quote, ForeignItemFn};
+
+    #[test]
+    fn test_generate_cpp_invokables_cfg() {
+        let method: ForeignItemFn = parse_quote! {
+            #[cfg(test_cfg_disabled)]
+            #[cxx_name = "voidInvokable"]
+            fn void_invokable(self: &MyObject);
+        };
+        let invokables = vec![ParsedMethod::mock_qinvokable(&method)];
+        let type_names = TypeNames::mock();
+        let opt = GeneratedOpt {
+            cfg_evaluator: Box::new(CfgEvaluatorTest::default()),
+        };
+        let generated =
+            generate_cpp_methods(&invokables.iter().collect(), &type_names, &opt).unwrap();
+
+        assert!(generated.methods.is_empty());
+    }
 
     #[test]
     fn test_generate_cpp_invokables() {
@@ -122,7 +151,12 @@ mod tests {
         let mut type_names = TypeNames::mock();
         type_names.mock_insert("QColor", None, None, None);
 
-        let generated = generate_cpp_methods(&invokables.iter().collect(), &type_names).unwrap();
+        let generated = generate_cpp_methods(
+            &invokables.iter().collect(),
+            &type_names,
+            &GeneratedOpt::default(),
+        )
+        .unwrap();
 
         // methods
         assert_eq!(generated.methods.len(), 5);
@@ -168,7 +202,8 @@ mod tests {
         type_names.mock_insert("A", None, Some("A1"), None);
         type_names.mock_insert("B", None, Some("B2"), None);
 
-        let generated = generate_cpp_methods(&invokables, &type_names).unwrap();
+        let generated =
+            generate_cpp_methods(&invokables, &type_names, &GeneratedOpt::default()).unwrap();
 
         // methods
         assert_eq!(generated.methods.len(), 1);