+++ crates/cxx-qt-gen/src/parser/externqobject.rs
@@ -3,7 +3,7 @@
 //
 // SPDX-License-Identifier: MIT OR Apache-2.0
 use crate::naming::Name;
-use crate::parser::CaseConversion;
+use crate::parser::{require_attributes, CaseConversion};
 use syn::{ForeignItemType, Ident, Result};
 
 /// A representation of a QObject to be generated in an extern C++ block
@@ -15,11 +15,23 @@ pub struct ParsedExternQObject {
 }
 
 impl ParsedExternQObject {
+    const ALLOWED_ATTRS: [&'static str; 6] = [
+        "cxx_name",
+        "rust_name",
+        "namespace",
+        "cfg",
+        "doc",
+        "qobject",
+        // TODO: support base, qproperty etc here?
+    ];
+
     pub fn parse(
         ty: ForeignItemType,
         module_ident: &Ident,
         parent_namespace: Option<&str>,
     ) -> Result<ParsedExternQObject> {
+        require_attributes(&ty.attrs, &Self::ALLOWED_ATTRS)?;
+
         Ok(Self {
             name: Name::from_ident_and_attrs(
                 &ty.ident,